+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [Unreleased]
 ## 0.1.0 - 2016-11-28
 ### Added
 - Created Leiningen project
+- Tests
+- Basic implementation of `derived` and `defderived` using an atom for cache
+- README
 
-[Unreleased]: https://github.com/your-name/defderived/compare/0.1.1...HEAD
-[0.1.1]: https://github.com/your-name/defderived/compare/0.1.0...0.1.1
+[Unreleased]: https://github.com/your-name/defderived/compare/0.1.0...HEAD